requirements.txt
# Core AI & ML
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0
tokenizers>=0.15.0
sentence-transformers>=2.2.2

# Speech & Audio
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
pygame>=2.5.2

# Web & Networking
requests>=2.31.0
beautifulsoup4>=4.12.2
wikipedia>=1.4.0
selenium>=4.15.0
urllib3>=1.26.18

# Utilities
python-dotenv>=1.0.0
pyautogui>=0.9.54
psutil>=5.9.5

# Code Analysis
astroid>=3.0.1
pylint>=3.0.2
black>=23.9.1

# Development
pytest>=7.4.2
pytest-asyncio>=0.21.1

# Additional dependencies for Ubuntu 25.10
numpy>=1.24.0
tqdm>=4.65.0
packaging>=21.0
